Flask/requirements.txt
absl-py==0.13.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
flatbuffers==1.12
gast==0.4.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
itsdangerous==2.0.1
imutils==0.5.4
Jinja2==3.0.1
json5==0.9.5
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.19.2
oauthlib==3.1.1
opencv-python==4.5.2.52
opt-einsum==3.3.0
Pillow==8.3.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
scikit-image==0.14.0
scikit-learn==0.20.4
scipy==1.2.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1